Abstract Database testing out of PgsqlAdapterTest
Much of the testing being done in PgsqlAdapterTest is
not specific to Postgres and will be able to be re-used
for testing the TestingAdapter.
